+++ CHANGELOG.md
@@ -23,8 +23,9 @@ A breaking change will get clearly marked in this log.
 ### Breaking changes
 
 - The response from `/trades` endpoint now can now contain two different kinds of trades, `orderbook` and `liquidity_pool` (new), which brought a few breaking changes ([#685](https://github.com/stellar/js-stellar-sdk/pull/685)):
-  - Some previously few mandatory fields were made into optional: `counter_offer_id`, `base_offer_id`, `base_account`, `counter_account`.
+  - Some previously few mandatory fields were made into optional. `counter_offer_id`, `base_offer_id` will only show up in orderbook trades while only one of `base_account` and `counter_account` will show in liquidity pool trades.
   - The `price` field changed from `{n: number; d: number;}` to `{n: string; d: string;}`.
+  - The links to "base" and "counter" returned from horizon can now point to either an account or a liquidity pool.
 
 ### Fix
 - Updated various developer dependencies to secure versions ([#671](https://github.com/stellar/js-stellar-sdk/pull/671)).